listring (shift-click) for hives and extractor
listring doesn't handle conditional movements, which is required for bees.  In other words, if you shift-click on an empty bottle it should load it into the extractor... but also if you shift-click on a a full frame it should load it into a different slot.  Listring can only do one or the other, which is why I needed to merge inventory lists and manually readjust which item goes into which position.