Pipeworks support
should be considered experimental.  I now believe that the pipeworks stuff that was there before was not fully implemented, and was not functional.
Added pipeworks for both extractor and artificial hive (Google "Flow Hive" to get an idea why).
Seems to work, but the pipeworks documentation is severely lacking, so I'm not sure what edge cases I might be missing.