Reduce rope selection box
Reduces the rope selection box to the actual rope itself.
Also for the ropebox, and knot at the bottom of the rope.

Greatly reduces the frequency of rope interfering with swords, picks, bows, placing nodes, opening chests, etc.  at the cost of making it somewhat harder to deliberately cut the rope with an ax (apx as hard as targeting a fence segment).