Update init.lua - remove gas nodes that stray too far from origin
I discovered a flaw in the `shut_down_lava_adjacent` algorithm. A gas seep may be further away from lava than the minimum distance, and the gas it produces will still eventually reach the lava if its downward path is unimpeded (e.g. the seeps are deep inside a cavity above the lava). The constant TNT explosions have a noticeable effect on FPS.

To get around this I am introducing a max distance limit on gas nodes from their point of origin, and if they cross it they'll be deleted.